Create Code
This code defines a variable gasProduction which is initially set to 0. It also defines a function trackGasProduction() which takes an argument amount and adds it to the gasProduction variable. Each time the function is called, it logs the updated gasProduction value to the console. You can call the function multiple times, passing in the desired amount of natural gas produced, and the function will add it up and log the total.